+++ http/api_versions.lua
@@ -1,0 +1,5 @@
+return {
+    V1 = 1,
+    V2 = 2,
+    UNKNOWN = 3,
+}